requirements.txt
Adafruit-Blinka==8.56.0
adafruit-circuitpython-busdevice==5.2.11
adafruit-circuitpython-connectionmanager==3.1.3
adafruit-circuitpython-lsm6ds==4.5.14
adafruit-circuitpython-register==1.10.2
adafruit-circuitpython-requests==4.1.10
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.77.0
Adafruit-PureIO==1.1.11
bidict==0.23.1
binho-host-adapter==0.1.6
lgpio==0.2.2.0
msgpack==1.1.0
packaging==24.2
pyftdi==0.56.0
pyserial==3.5
python-can==4.5.0
RPi.GPIO==0.7.1
rpi_ws281x==5.0.0
six==1.17.0
spidev==3.6
sysv_ipc==1.1.0
transitions==0.9.2
typing_extensions==4.13.2
wrapt==1.17.2
